Исправления замечаний по коду
