+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
             android:name=".about.AboutActivity"
             android:exported="false" />
         <activity
-            android:name=".main.MainActivity"
+            android:name=".main.activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
