+++ app/src/main/java/com/a4nt0n64r/notes/MVPInterfaces.kt
@@ -7,9 +7,9 @@ import com.a4nt0n64r.notes.main.models.NoteUI
  */
 interface MainActivityInterface {
     /**
-     *  Показ Snackbar-а
+     *  Показ Snackbar
      *
-     *  @param message - сообщение
+     *  @param message сообщение
      */
     fun showSnackBar(message: String)
 
@@ -26,7 +26,7 @@ interface MainPresenter {
     /**
      *  Обработка клика на кнопку "Save"
      *
-     *  @param message - сообщение
+     *  @param message сообщение
      */
     fun saveClicked(message: String)
 }
@@ -38,7 +38,7 @@ interface ListPresenter {
     /**
      *  Обработка клика на заметку
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun noteClicked(note: NoteUI)
 
@@ -55,14 +55,14 @@ interface ListView {
     /**
      *  Показать экран с информацией о заметке
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun navigateToNoteInfo(note: NoteUI)
 
     /**
      *  Заполнение RecyclerView списком заметок
      *
-     *  @param listOfNotes - список заметок
+     *  @param listOfNotes список заметок
      */
     fun setupNotesList(listOfNotes: List<NoteUI>)
 }
@@ -79,21 +79,21 @@ interface NotePresenter {
     /**
      *  Установка заметки в презентер
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun setNote(note: NoteUI)
 
     /**
      *  Обработка клика на кнопку "Search"
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun searchClicked(note: NoteUI)
 
     /**
      *  Обработка клика на кнопку "Save"
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun saveClicked(note: NoteUI)
 }
@@ -110,22 +110,22 @@ interface NoteView {
     /**
      *  Отображение заметки
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun showNote(note: NoteUI)
 
     /**
      *  Сохранение заметки
      *
-     *  @param note - заметка
+     *  @param note заметка
      */
     fun saveNote(note: NoteUI)
 
     /**
-     *  Пробуем поискать заметку на YouTube, если не выходит - показываем Snackbar с сообщением
+     *  Пробуем поискать заметку на YouTube, если не выходит показываем Snackbar с сообщением
      *  ошибки
      *
-     *  @param query - Текст для поиска
+     *  @param query текст для поиска
      */
     fun trySendYoutubeIntent(query: String)
 }
