+++ app/src/main/java/com/a4nt0n64r/notes/extentions/Extentions.kt
@@ -1,0 +1,19 @@
+package com.a4nt0n64r.notes.extentions
+
+import android.app.Activity
+import android.content.Context
+import android.view.View
+import android.view.inputmethod.InputMethodManager
+
+/**
+ * Функция скрывания клавиатуры
+ */
+fun Activity.hideKeyboard() {
+    hideKeyboard(currentFocus ?: View(this))
+}
+
+private fun Context.hideKeyboard(view: View) {
+    val inputMethodManager =
+        getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
+    inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
+}
