+++ app/src/main/java/com/a4nt0n64r/notes/main/activity/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.a4nt0n64r.notes.main
+package com.a4nt0n64r.notes.main.activity
 
 import android.app.Activity
 import android.content.Context
@@ -10,6 +10,7 @@ import androidx.appcompat.app.AppCompatActivity
 import com.a4nt0n64r.notes.*
 import com.a4nt0n64r.notes.about.AboutActivity
 import com.a4nt0n64r.notes.databinding.ActivityMainBinding
+import com.a4nt0n64r.notes.extentions.hideKeyboard
 import com.a4nt0n64r.notes.main.fragments.list.ListFragment
 import com.a4nt0n64r.notes.main.fragments.note.NoteFragment
 import com.google.android.material.snackbar.Snackbar
@@ -38,33 +39,13 @@ class MainActivity : AppCompatActivity(), MainActivityInterface {
         Snackbar.make(binding.constraint, message, Snackbar.LENGTH_LONG).show()
     }
 
-    private fun openAboutActivity() {
-        startActivity(Intent(this, AboutActivity::class.java))
-    }
 
-    override fun onResume() {
-        super.onResume()
-
-        binding.about.setOnClickListener {
-            openAboutActivity()
-        }
-    }
-
-    private fun Activity.hideKeyboard() {
-        hideKeyboard(currentFocus ?: View(this))
-    }
-
-    private fun Context.hideKeyboard(view: View) {
-        val inputMethodManager =
-            getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
-        inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
-    }
 
     /**
      * Функция для навигации между фрагментами
      *
-     * @param destination - строковая константа для фрагмента-назначения
-     * @param bundle - Бандл
+     * @param destination строковая константа для фрагмента-назначения
+     * @param bundle Бандл
      */
     fun navigate(destination: String, bundle: Bundle?) {
         when (destination) {
