+++ app/src/main/java/com/a4nt0n64r/notes/main/activity/MainPresenterImpl.kt
@@ -1,4 +1,4 @@
-package com.a4nt0n64r.notes.main
+package com.a4nt0n64r.notes.main.activity
 
 import com.a4nt0n64r.notes.MainActivityInterface
 import com.a4nt0n64r.notes.MainPresenter
@@ -6,7 +6,7 @@ import com.a4nt0n64r.notes.MainPresenter
 /**
  * Презентер
  *
- * @param view  - Активити
+ * @param view Активити
  */
 class MainPresenterImpl(private val view: MainActivityInterface) : MainPresenter {
     override fun saveClicked(message:String) {
