+++ app/src/main/java/com/a4nt0n64r/notes/main/fragments/list/ListFragment.kt
@@ -1,16 +1,16 @@
 package com.a4nt0n64r.notes.main.fragments.list
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.res.ResourcesCompat
 import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.a4nt0n64r.notes.*
+import com.a4nt0n64r.notes.about.AboutActivity
 import com.a4nt0n64r.notes.databinding.FragmentListBinding
-import com.a4nt0n64r.notes.main.MainActivity
+import com.a4nt0n64r.notes.main.activity.MainActivity
 import com.a4nt0n64r.notes.main.models.NoteUI
 
 /**
@@ -18,9 +18,10 @@ import com.a4nt0n64r.notes.main.models.NoteUI
  */
 class ListFragment : Fragment(), ListView {
 
-    lateinit var binding: FragmentListBinding
+    private var _binding: FragmentListBinding? = null
+    private val binding get() = _binding!!
 
-    private lateinit var presenter: ListPresenter
+    private var presenter: ListPresenter? = null
 
     lateinit var adapter: NotesAdapter
 
@@ -28,7 +29,7 @@ class ListFragment : Fragment(), ListView {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentListBinding.inflate(layoutInflater)
+        _binding = FragmentListBinding.inflate(layoutInflater)
         return binding.root
     }
 
@@ -38,11 +39,21 @@ class ListFragment : Fragment(), ListView {
         presenter = ListPresenterImpl(this)
 
         setupRecyclerView()
-        presenter.getNotes()
+        presenter?.getNotes()
 
         binding.noteList.addItemDecoration(
             MarginItemDecoration(resources.getDimensionPixelSize(R.dimen.app_margin_8dp))
         )
+
+        binding.about.setOnClickListener {
+            openAboutActivity()
+        }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+        presenter = null
     }
 
     private fun setupRecyclerView() {
@@ -56,7 +67,11 @@ class ListFragment : Fragment(), ListView {
     }
 
     private fun noteClick(note: NoteUI) {
-        presenter.noteClicked(note)
+        presenter?.noteClicked(note)
+    }
+
+    private fun openAboutActivity() {
+        startActivity(Intent(requireContext(), AboutActivity::class.java))
     }
 
     override fun navigateToNoteInfo(note: NoteUI) {
